Add rustls-native-roots feature
In enterprise settings, it is common to require using corporate
certificates.

This allows for enabling the reqwest `rustls-tls-native-roots` feature,
which fixes this on Linux hosts (Windows is a bit more messy, but can
work).

I'm open to another name, but reqwest uses `rustls-tls-native-roots`,
but the dependency that is enabled is called `rustls-native-certs`...